requirements.txt
# Core Dependencies
python-dotenv==1.0.0
neo4j==5.14.0

# AWS
boto3==1.34.0

# OpenAI
openai==1.12.0

# PDF Processing
PyMuPDF==1.23.8
camelot-py[cv]==0.11.0
opencv-python==4.9.0.80

# NLP and Text Processing
sentence-transformers==2.3.1
thefuzz==0.20.0
python-Levenshtein==0.23.0

# Vector Search
faiss-cpu==1.7.4

# Web Framework
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3

# Utilities
pandas==2.1.4
numpy==1.26.3
tqdm==4.66.1
requests==2.31.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0